+++ cocos2d/cocos/renderer/backend/opengl/DeviceGL.cpp
@@ -64,7 +64,7 @@ CommandBuffer* DeviceGL::newCommandBuffer()
     return new (std::nothrow) CommandBufferGL();
 }
 
-Buffer* DeviceGL::newBuffer(unsigned int size, BufferType type, BufferUsage usage)
+Buffer* DeviceGL::newBuffer(unsigned long size, BufferType type, BufferUsage usage)
 {
     return new (std::nothrow) BufferGL(size, type, usage);
 }
